requirements.txt
fastapi
uvicorn
bs4
requests
langchain
langchain-community
unstructured
libmagic
lxml
torch
python-dotenv
huggingface_hub
transformers
langchain-huggingface
langchain_experimental
ggshield
streamlit